classify: fix classify filter trace del cli processing
When a 'del' is used to delete a classify table, only the
mask is needed to locate the table.  Any match vector is
unneeded.  The tests failed to notice this, but if the
test is run by hand in vppctl, it issues a parse error.

Fix the test so that it doesn't supply irrelevant data.
Fix the CLI processing to read always complete newline
terminated line of input instead.  This allows unneeded
CLI parameters to be ignored.  It also necessitated
fixing a trace test which had then erroneously split
a single CLI command over multiple lines.

While in the area, fix a latent bug on table matching
where a test for compatible mask vector sizes were
not matching impedance properly (byte vs ux32x4).

Type: fix
Signed-off-by: Jon Loeliger <jdl@netgate.com>
Change-Id: I1177ab1dd417f3d11f30eecbaa2b0fb1015c3ab5