vrrp: increase stack size of process node
Type: fix

The process node which wakes up when a timer expires and transitions
a backup node to master state may call a function to add a MAC address
to an interface. This works fine for some devices, but with DPDK 20.11
on i40e interfaces, the i40e PMD functions which enact the change cause
the stack to be exhausted. Increase the stack size for the node.

Change-Id: I824603e162f4f6d680486706210986572f0d9845
Signed-off-by: Matthew Smith <mgsmith@netgate.com>